Build workers in same linking step as explore
Also move Vite config to use async instead of sync (which is not recommended as it is blocking)

Explore now depends on the workers project. This way any shared code between explore and workers is only linked to js once, instead of for each module. I _think_ this should save on bundle size.

We _could_ also improve the build speed by not making workers a direct dependency of explore, but instead have a separate project that depends on both explore and workers, and then build that project. But that is probably future work. Workers is also not _that_ big, so splitting up Explore would probably yield better gains. The project graph would be more diamond-shaped:

```
          ┌─────► explore────┐
          │                  │
          │                  │
common, etc                  ├──────►webapp
          │                  │
          │                  │
          └─────► workers────┘
```