Fix observations not appearing in groups on first render
Because the `existence` was an `Option` in two different query elements, the `GroupObs` existence was not being checked correctly on group children, as it would always be `None` on those.
With this change the `elements` of groups now only query the `id`, and existence is checked on the 'root' elements of the group tree.